pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "circuit-pipeline"
version = "0.1.0"
description = "Neural circuit analysis pipeline"
requires-python = ">=3.10"

dependencies = [
  "numpy",
  "scipy",
  "matplotlib",
  "pyyaml"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
